requirements.txt
django==1.11.7
confusable-homoglyphs==3.0.0
django-registration==2.3
mysqlclient==1.3.12
django-assets==0.12
pytest-django==3.1.2
pytest-cov==2.5.1
django-honeypot==0.6.0
django-hosts==3.0
django-crispy-forms==1.7.0
django-debug-toolbar==1.9.1
geopy==1.11.0
django-argonauts==1.2.0
https://github.com/philippbosch/django-geoposition/archive/django-1.11.zip
tablib==0.12.1
djangorestframework==3.7.7
social-auth-app-django==2.1.0
